Now using a default bundle for scoop and filter
Scoop and filter code had been made dependent on a bundle_id
attribute being present on a patient. This makes sense for Cypress
but not so much for applications like popHealth. For now, if there
is no bundle_id present on the patient, the code will select the
latest bundle available for its code sets. In the future, we should
probably decouple this further.