Cleaning up the measure model and associated measure fixtures
Moving over some code from Cypress that we needed to extract measure
information in the QRDA Category I scoop and filter logic.

Smoking gun logic has been removed. This will be moved to Cypress.

The fixtures for the measures were replaced with the fixtures from
Cypress. These appear to be closer to the JSON structure that is
expected by most of the code for dealing with measures. However,
I'm still finding inconsistencies in these fixtures.