Starting work on prefilters for the CQM calculation M/R jobs
This code attempts to look at the measure structure to come up with
a MongoDB query to prefilter records before doing measure calculation.
Currently, this will only work for measures with birthdate criteria
specified in the IPP relative to the measure period. This will require
work in QME to take advantage of the prefilters.