Implementing fulfills operator parsing in hqmf r2.
Added Reference class to the models  to allow Entries to have arbitary
references to other entries.  References are made up of a type (what
kind of reference it is such as fulfills) and a referenced_type and
referenced_id which are the Entry class type and mongo id of the entry itself

Mongoid references do not work on embedded objects so they could not be used.

Added additional TypedReference class to hqmf library to allow capturing
the fulfills reference .