Return performance rate validation to not rounding expected values, and add tests
Problem: Performance Rates (according to the spec) should NOT be rounded to 6 decimal places, but rather truncated. HDS's Performance Rate Valdiator was changed to round them in a previous pull request.

Solution: Change the code back to the way it was before that pull request, and add comments to the performance rate valdiator talking about why we're not rounding the performance rate validator. Also, add tests to ensure that we're correctly validating repeating decimals such as 2/3