Now handles specific occurrences without errors.
Previously, specific occurrences were causing errors when parsing
due to lack of occurences_map.  This has been modified, along with
how the criteria are parsed (obtained once from the xml) and how
the occurrences are tested (the names are matched for easier
comparison)

Handles erroneous birth dates and stratifications

Removes stratification property from population in comparison (it
seems it is present in HQMF, but not SimpleXML), removes the
birthdate from a specific measure (age is not used), and cleans up
the birthdate from data criteria in SimpleXML (when there are already
temporally restricted birthdate elements)

Changed data criteria filtering and measure test corrections

Data filtering was removing elements that should have been kept,
this is now closer to what it should be taking out and is much
easier to change. The test has been corrected to handle denominator
exclusions of any name that are unnecessary, and fixed a bug with
the individual measure correction CMS82v3.

Small test correction for ORDINALITY -> ORDINAL

The test only handles one measure correction (CMS26v3), more can be
added as necessary, or can be made general if necessary. Also
prevents error occurring due to titleize being called on a nil
object.