Added base comments to all "to_models", changed do_not_group
Do not group has been changed to "is_derived_specific_occurrence_variable", and
it's handler has been changed similarly. This needs to be handle during the
extract grouper methods, as there may be a case where the handler should use the
GROUP version of it's reference criteria rather than the regular.